web-app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.9
google-generativeai==0.3.2
python-dotenv==1.0.1
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

tensorflow>=2.15.0
